Lab 6/requirements.txt
google-cloud-pubsub
google-cloud-storage